% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipca2.R
\name{biocIPCA}
\alias{biocIPCA}
\alias{biocPCA}
\alias{biocIPCA_chunked}
\alias{biocPCA_chunked}
\title{use sklearn.incrementalPCA interfaced to package-local python module}
\usage{
biocIPCA(mat, n_comp, batch_size)

biocPCA(mat, n_comp)

biocIPCA_chunked(mat, chunk_size, n_comp, batch_size)
}
\arguments{
\item{n_comp}{number of PCs to compute}

\item{batch_size}{number of records handled in each batch}

\item{R}{matrix}
}
\description{
use sklearn.incrementalPCA interfaced to package-local python module
}
\note{
biocIPCA_chunked uses partial_fit over all chunks.
}
\examples{
# small example
dd = data.matrix(iris[,1:4])
Rpcs = prcomp(dd)
Ppcs = biocIPCA(dd, 4, 10)
cor(Rpcs$x, Ppcs$rotated)
# bigger
data(bano5k)
bpc = prcomp(bano5k)
Pbpc = biocIPCA(bano5k, 5, 1000) # need mem profiling
cor(bpc$x[,1:5], Pbpc$rotated)
Pbpc[[1]]$explained_variance_
Pbpc2 = biocIPCA(bano5k, 5, 5000) # may have higher mem cost
cor(bpc$x[,1:5], Pbpc2$rotated) # better
Pbpc2[[1]]$explained_variance_
}
