% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_hdf5.R
\name{combine_files}
\alias{combine_files}
\title{combine hdf5 datasets from multiple files into a new hdf5 file}
\usage{
combine_files(tenxfiles)
}
\arguments{
\item{tenxfiles}{collection of tenxfiles}
}
\description{
combine hdf5 datasets from multiple files into a new hdf5 file
}
