% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoret.R
\name{demArr}
\alias{demArr}
\title{getting a constant array from python to R}
\usage{
demArr(size = 25L, dims = c(4L, 3L, 2L))
}
\arguments{
\item{size}{integer giving overall python size of array}

\item{dims}{array dimensions in python idiomatic order}
}
\description{
getting a constant array from python to R
}
