% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfiles.R
\name{getfiles}
\alias{getfiles}
\title{get hdf5 files from working directory}
\usage{
getfiles(filepath)
}
\arguments{
\item{filepath}{path to the directory having the hdf5 files}
}
\description{
get hdf5 files from working directory
}
