% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incremental_pca.R
\name{incremental_pca}
\alias{incremental_pca}
\title{perform incremental pca on a hdf5 file}
\usage{
incremental_pca(filepath)
}
\arguments{
\item{filepath}{path to hdf5 file}
}
\description{
perform incremental pca on a hdf5 file
}
