% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipca.R
\docType{methods}
\name{ipca,H5pymat,integer,integer,logical-method}
\alias{ipca,H5pymat,integer,integer,logical-method}
\alias{ipca}
\title{use chunks from remote HDF5 source to compute incremental PCA}
\usage{
\S4method{ipca}{H5pymat,integer,integer,logical}(src, ncomp, batch_size,
  transpose = TRUE)
}
\arguments{
\item{src}{H5pymat instance}

\item{ncomp}{number of PCs to compute}

\item{batch_size}{number of records per batch}

\item{transpose}{logical}
}
\description{
use chunks from remote HDF5 source to compute incremental PCA
}
\examples{
d = data.matrix(iris[,1:4])
p = prcomp(d)
dim(p$x)
h5targ = tempfile()
require(rhdf5)
h5createFile(h5targ)
h5write(d, h5targ, "iris")
hmat = H5pymat(h5targ)
tst = ipca(hmat, ncomp=4L, batch_size=50L, transpose=TRUE)
dim(tst)
}
